+++ libvast/include/vast/type.hpp
@@ -460,7 +460,7 @@ public:
   friend std::span<const std::byte> as_bytes(const int64_type&) noexcept;
 
   /// Constructs data from the type.
-  [[nodiscard]] static integer construct() noexcept;
+  [[nodiscard]] static int64_t construct() noexcept;
 
   /// Converts the type into an Arrow DataType.
   [[nodiscard]] static std::shared_ptr<arrow_type> to_arrow_type() noexcept;
