+++ libvast/test/format/json.cpp
@@ -89,7 +89,7 @@ TEST(json to data) {
     std::integral_constant<
       int, caf::detail::tl_index_of<data::types, double>::value>());
   CHECK(std::abs(r - double{4.2}) < 0.000001);
-  CHECK_EQUAL(materialize(slice.at(0, 3)), integer{-1337});
+  CHECK_EQUAL(materialize(slice.at(0, 3)), int64_t{-1337});
   CHECK_EQUAL(materialize(slice.at(0, 4)), "0123456789®\r\n");
   CHECK_EQUAL(materialize(slice.at(0, 5)), "42.42");
   std::array<std::uint8_t, 4> addr1{147, 32, 84, 165};
