+++ libvast_test/src/table_slices.cpp
@@ -265,7 +265,7 @@ void table_slices::test_append_column_to_index() {
   slice.append_column_to_index(1, *idx);
   CHECK_EQUAL(idx->offset(), 2u);
   constexpr auto less = relational_operator::less;
-  CHECK_EQUAL(unbox(idx->lookup(less, make_view(integer{3}))), make_ids({1}));
+  CHECK_EQUAL(unbox(idx->lookup(less, make_view(int64_t{3}))), make_ids({1}));
 }
 
 } // namespace fixtures
